requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
bleach==6.0.0
boto3==1.28.31
botocore==1.31.31
dj-database-url==2.1.0
Django==4.2.3
django-bootstrap-v5==1.0.11
django-dotenv==1.4.2
django-on-heroku==1.1.2
django-recaptcha==3.0.0
django-storages==1.13.2
django-taggit==4.0.0
gunicorn==21.2.0
jmespath==1.0.1
Markdown==3.4.4
packaging==23.1
Pillow==10.0.0
psycopg2-binary==2.9.7
pycountry==22.3.5
python-dateutil==2.8.2
s3transfer==0.6.2
shortuuid==1.0.11
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tinycss2==1.1.1
typing_extensions==4.7.1
Unidecode==1.3.6
urllib3==1.26.16
webencodings==0.5.1
whitenoise==6.5.0
